Add return type builtin checks using reflection
Using the ReflectionType instead of the custom BuiltinType class makes
sure that future builtin types will be supported and also solves the
current issue related to the fact that "void" is not a builtin type in
PHP prior to 7.3.
We also had to make some small changes for the use of the special return
or parameter value "self".